requirements.txt
# Core
openai>=1.12.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
slowapi>=0.1.9

# Database (PostgreSQL + pgvector)
psycopg2-binary>=2.9.9
pgvector>=0.3.0
numpy>=1.24.0

# Analytics / Visualization
matplotlib>=3.7.0

# Scraping
tqdm>=4.66.0
playwright>=1.40.0

# Document parsing
python-docx>=1.1.0
python-multipart>=0.0.9

# Resilience / retry
tenacity>=8.2.0

# Evaluation (RAG metrics)
ragas>=0.1.0
datasets>=2.14.0
langchain-openai>=0.1.0
eval_type_backport>=0.1.0